Support template-haskell and tests on GHC < 8.8
We can effortlessly support GHC 8.0+ instead of GHC 8.8+ for both
building and testing, so let's do that.

GHC's before 8.8 (base-4.13) don't export `MonadFail` to Prelude.

Relax the bounds on `template-haskell` to support GHC 8.0+

Tighten the bounds on `containers` and `ansi-terminal` to exclude
unsupported versions.

Drop bounds specification in the test suite for dependencies shared with
the library; the library already constrains these dependencies and it
reduces maintenance burden.